There's some sections of code commented out because it requires other classes to run properly. Also, there is no place to store the file locations because such variables were not specified in client Profile, nor did the spreadsheet dictate their creation.
Also, instead of greying out the register button hitting the register
button without entering all the information creates a error popup
because I was not sure how to make the button unclickable and have it
still be a JButton.

There was not a writeRegistration method in client profile either, so
that part of the specifications was temporarily disregarded.